Fix compatibility issue with Windows and improve code quality
