--- src/main/java/Launcher.java
@@ -1,8 +1,0 @@
-import duke.Main;
-import javafx.application.Application;
-
-public class Launcher {
-    public static void main(String[] args) {
-        Application.launch(Main.class, args);
-    }
-}
