+++ src/main/java/duke/Main.java
@@ -1,33 +1,9 @@
 package duke;
 
-import java.io.IOException;
-
 import javafx.application.Application;
-import javafx.fxml.FXMLLoader;
-import javafx.scene.Scene;
-import javafx.scene.layout.AnchorPane;
-import javafx.stage.Stage;
 
-/**
- * A GUI for Duke using FXML.
- */
-public class Main extends Application {
-
-    private Duke duke = new Duke();
-
-    @Override
-    public void start(Stage stage) {
-        try {
-            duke.load();
-            FXMLLoader fxmlLoader = new FXMLLoader(Main.class.getResource("/view/MainWindow.fxml"));
-            AnchorPane ap = fxmlLoader.load();
-            Scene scene = new Scene(ap);
-            stage.setScene(scene);
-            fxmlLoader.<MainWindow>getController().setDuke(duke);
-            stage.show();
-            fxmlLoader.<MainWindow>getController().greetMessage();
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
+public class Main {
+    public static void main(String[] args) {
+        Application.launch(Launcher.class, args);
     }
 }
